refactor: change variable type from BigDecimal to Integer

+++ src/main/java/com/deathz/laborcalc/domain/model/MonthlyCompetenceDetail.java
@@ -8,7 +8,7 @@ public record MonthlyCompetenceDetail(
     LocalDate periodEnd,
     Integer daysWorked,
     BigDecimal currentMinimumWage,
-    BigDecimal additionalPercentage,
+    Integer additionalPercentage,
     BigDecimal fullAdditionalAmount,
     BigDecimal proportionalAdditionalAmount,
     BigDecimal thirteenthSalaryProportion,
